\name{datax}
\docType{data}
\alias{datax}
\title{Examplary dataset containing parameter settings}
\description{dataset coming from a sheet metal forming process}
\usage{data(datax)}
\format{Table containing parameter settings, X1 and X2}
\references{
Sonja Kuhnt and Martina Erdbruegge (2004). A strategy of robust paramater design for multiple responses, 
Statistical Modelling; 4: 249-264, TU Dortmund.
}
\keyword{datasets}
