\name{datay}
\docType{data}
\alias{datay}
\title{Examplary dataset containing Responses}
\description{dataset coming from a sheet metal forming process}
\usage{data(datay)}
\format{Table containing responses, Y1 and Y2.}
\references{
Sonja Kuhnt and Martina Erdbruegge (2004). A strategy of robust paramater design for multiple responses, 
Statistical Modelling; 4: 249-264, TU Dortmund.
}
\keyword{datasets}
