\name{JOP-package}
\alias{JOP-package}
\docType{package}
\title{
JOP: Joint Optimization Plot
}
\description{
Consider for example the development of a production process where a set of response variables with
corresponding target values depends on several control parameters. Here an optimal 
parameter setting is of great interest in order to get a high quality product.
 
JOP calculates optimal design parameters associated with a given sequence of weight matrices
based on the minimization of a risk function introduced by Pignatiello (1993). Furthermore JOP 
visualizes the optimal design parameters and the appropriate predicted responses using the joint 
optimization plot introduced by Kuhnt and Erdbruegge (2004). 
}
\details{
\tabular{ll}{
Package: \tab JOP\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2010-05-07\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{
Sonja Kuhnt and Nikolaus Rudak <rudak@statistik.tu-dortmund.de>.
}
\references{
Sonja Kuhnt and Martina Erdbruegge (2004), "A strategy of robust paramater design for multiple responses", 
Statistical Modelling; 4: 249-264, TU Dortmund.

Joseph J. Pignatiello (1993), "Strategies for robust multiresponse quality engineering", IIE Transactions 25, 5-15, Texas A&M University.
}